style(directory) RoomManager.java renamed and change directory
