+++ src/main/java/com/souma1024/shogiv2/controller/rest/RoomController.java
@@ -15,8 +15,8 @@ import com.souma1024.shogiv2.dto.room.JoinRoomResqponse;
 import com.souma1024.shogiv2.entity.Room;
 import com.souma1024.shogiv2.enums.common.PlayerSide;
 import com.souma1024.shogiv2.repository.RoomRepository;
+import com.souma1024.shogiv2.service.RoomSessionManager;
 import com.souma1024.shogiv2.service.RoomService;
-import com.souma1024.shogiv2.websocket.RoomManager;
 
 @RestController
 @RequestMapping("/api")
@@ -44,7 +44,7 @@ public class RoomController {
         Room room = roomService.createRoom(timeLimit);
 
         CreateRoomResponse response = new CreateRoomResponse(room.getRoomId(), room.getFirstPlayerId(), timeLimit, room.getStatus(), "ルーム作成に成功しました");
-        RoomManager.getInstance().canAddPlayer(response.getRoomId(), new Player(response.getPlayerId(), PlayerSide.SENTE));
+        RoomSessionManager.getInstance().canAddPlayer(response.getRoomId(), new Player(response.getPlayerId(), PlayerSide.SENTE));
         return ResponseEntity.status(201).body(response);
     }
 
@@ -53,7 +53,7 @@ public class RoomController {
         try {
             JoinRoomResqponse response = roomService.joinRoom(roomId);
             Player player = new Player(response.getPlayerId(), PlayerSide.GOTE); // or SENTE
-            RoomManager.getInstance().canAddPlayer(roomId, player);
+            RoomSessionManager.getInstance().canAddPlayer(roomId, player);
 
             return ResponseEntity.ok(response);
         } catch (IllegalArgumentException e) {
